CLOUD-487: support string:string maps in go-i18n
This flat file format assumes everything is a map of strings to objects. We want to be able to support string:string, since that's the default that Localise uses if there's only one translation. We also want to be able to support string:map, to allow specifying plural rules.

Additionally, update our fork to include proper module information, and fix a couple of tests that had go vet failures.