BACK-750: fix encoding of fallback translations
We want to use the fallback behaviour instead of storing thousands of identical strings (and instead of adding English strings to the French locale files). However, the fallback behaviour of go-i18n is buggy: it uses text/template for found translations, but html/template for fallbacks. This causes different encoding behaviour!

Use text/template instead so fallback behaviour matches the normal translation behaviour.